# apps/taptrade-platform/go-platform/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tap Trade API
  version: "1.0.0"
  description: >
    Tap Trade prediction-market gateway API for non-redeemable gameplay points.
    Binary YES/NO outcomes use point-cents subunits on a 0-100 implied
    probability scale; settled winning shares return 100 point-cents per share.
    Points are gameplay-only units with no external value path.

    Coverage: this spec documents the core public, trading, portfolio,
    reward, loyalty, social, leaderboard, bot, auth proxy,
    admin account-review, admin discovery taxonomy, admin market operations,
    and admin market-lifecycle surfaces verified against the live handlers and
    DTOs as of 2026-06. CMS and remaining compatibility surfaces are
    intentionally out of scope here and tracked for documentation in
    IMPROVEMENT_PLAN.md P2-05 / G-04. The previous inherited spec is archived
    outside this launch document.
servers:
  - url: http://localhost:18080
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly session cookie issued by the auth service. Mutating requests also require the X-CSRF-Token header (double-submit cookie).
    BotApiKey:
      type: http
      scheme: bearer
      description: "Bot/partner API key: Authorization: Bearer tna_<prefix>_<secret>. Scopes: read, trade."
  schemas:
...
    KYCStatus:
      type: object
      properties:
        userId: { type: string }
        status: { type: string, enum: [unverified, pending, approved, declined, blocked] }
        riskLevel: { type: string }
        lastVerifiedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        documentsSubmitted: { type: array, items: { type: string } }
        rejectionReasons: { type: array, items: { type: string } }
    KYCDocument:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        type: { type: string }
        documentId: { type: string }
        issuingCountry: { type: string }
        expiryDate: { type: string }
        submittedAt: { type: string, format: date-time }
        verifiedAt: { type: string, format: date-time }
        status: { type: string, enum: [submitted, verifying, approved, rejected] }
        rejectReason: { type: string }
...
paths:
...
  /api/v1/compliance/geo/check:
    get:
      summary: Report whether a country (and optional state) is an approved jurisdiction
      parameters:
        - { name: country, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        "200":
          description: Approval decision for the location
          content:
            application/json:
              schema:
                type: object
                properties:
                  country: { type: string }
                  state: { type: string }
                  approved: { type: boolean }
...
  /api/v1/compliance/kyc/status:
    get:
      summary: Read the caller's identity-verification status
      description: Session-bound. A userId query is allowed only when it matches the session user.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: userId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Current KYC status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { $ref: "#/components/schemas/KYCStatus" }
  /api/v1/compliance/kyc/submit-document:
    post:
      summary: Submit an identity document for review
      description: Session-bound; the body's userId must match the session user. Stores document metadata, never the file.
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, type]
              properties:
                userId: { type: string }
                type: { type: string, description: passport, driver_license, id_card, … }
                documentId: { type: string }
                issuingCountry: { type: string }
                expiryDate: { type: string }
      responses:
        "201":
          description: Stored document metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  document: { $ref: "#/components/schemas/KYCDocument" }
  /api/v1/compliance/kyc/documents:
    get:
      summary: List the caller's submitted identity documents
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: userId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Submitted documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId: { type: string }
                  documents: { type: array, items: { $ref: "#/components/schemas/KYCDocument" } }
                  total: { type: integer }
...